ok K8829

+++ public/main/inc/lib/usergroup.lib.php
@@ -8,7 +8,10 @@ use Chamilo\CoreBundle\Enums\ActionIcon;
 use Chamilo\CoreBundle\Enums\ObjectIcon;
 use Chamilo\CoreBundle\Enums\ToolIcon;
 use Chamilo\CoreBundle\Framework\Container;
+use Chamilo\CoreBundle\Service\StandardizationService;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
+use Symfony\Polyfill\Intl\Normalizer\Normalizer;
+
 
 /**
  * Class UserGroup.
@@ -1235,6 +1238,36 @@ class UserGroupModel extends Model
                         }
                     }
 
+                    Database::delete(
+                        $this->usergroup_rel_course_table,
+                        [
+                            'usergroup_id = ? AND course_id = ?' => [
+                                $usergroup_id,
+                                $course_id,
+                            ],
+                        ]
+                    );
+                }
+                if (0 != $sessionId && 0 != $groupId) {
+                    $this->subscribe_sessions_to_usergroup($groupId, [0]);
+                } else {
+                    $s = $sessionId;
+                }
+            }
+        }
+    }
+
+    public function unsubscribe_only_courses_from_usergroup($usergroup_id, $delete_items, $sessionId = 0)
+    {
+        $sessionId = (int) $sessionId;
+        // Deleting items.
+        if (!empty($delete_items)) {
+            $user_list = $this->get_users_by_usergroup($usergroup_id);
+
+            $groupId = isset($_GET['id']) ? (int) $_GET['id'] : 0;
+            foreach ($delete_items as $course_id) {
+                $course_info = api_get_course_info_by_id($course_id);
+                if ($course_info) {
                     Database::delete(
                         $this->usergroup_rel_course_table,
                         [
@@ -3203,6 +3236,92 @@ class UserGroupModel extends Model
 
         return Database::store_result($result, 'ASSOC');
     }
+
+    public function getUsersInAndOutOfCourse($usergroupId, $courseId): array
+    {
+        $usergroupModel = new UserGroupModel();
+
+        $usersInUsergroup = $usergroupModel->get_users_by_usergroup($usergroupId);
+
+        $data = [];
+        $data['error'] = null;
+        $data['warning'] = null;
+        $data['usersSubscribedToCourse'] = [];
+        $data['usersNotSubscribedToCourse'] = [];
+
+        if (sizeof($usersInUsergroup) > 0) {
+
+            $userManager = new UserManager();
+            $userRepository = $userManager->getRepository();
+
+            $courseManager = new CourseManager();
+            $usersInCourse = $courseManager->get_user_list_from_course_code($courseId);
+
+            $usersSubscribedToCourse = [];
+            $usersNotSubscribedToCourse = [];
+            foreach ($usersInUsergroup as $userId) {
+                $user = $userRepository->find($userId);
+                if (array_key_exists($userId, $usersInCourse)) {
+                    $usersSubscribedToCourse[] = $user;
+                } else {
+                    $usersNotSubscribedToCourse[] = $user;
+                }
+            }
+
+            $data['usersSubscribedToCourse'] = StandardizationService::sortByNameByCountryAndStandardizeName($usersSubscribedToCourse, true);
+            $data['usersNotSubscribedToCourse'] = StandardizationService::sortByNameByCountryAndStandardizeName($usersNotSubscribedToCourse, true);;
+        } else {
+            $data['warning'] = get_lang('No user is subscribed to this class');
+        }
+
+        return $data;
+    }
+
+    public function getUsersAndCoursesSubscribedToAUserGroup($usergroupId):array
+    {
+        $data = [];
+        $data['error'] = null;
+        $data['warning'] = null;
+        $data['usersSubscribedToUsergroup'] = [];
+        $data['coursesSubscribedToUsergroup'] = [];
+
+        $usergroupLib = new UserGroupModel();
+
+        $userManager = new UserManager();
+        $userRepository = $userManager->getRepository();
+        $usersSubscribedToUsergroupIds = $usergroupLib->get_users_by_usergroup($usergroupId);
+        if (count($usersSubscribedToUsergroupIds) > 0) {
+            $usersSubscribedToUsergroup = [];
+            foreach ($usersSubscribedToUsergroupIds as $userId) {
+                $usersSubscribedToUsergroup[] = $userRepository->find($userId);
+            }
+            $data['usersSubscribedToUsergroup'] = StandardizationService::sortByNameByCountryAndStandardizeName($usersSubscribedToUsergroup, true);
+        }
+
+        $courseManager = new CourseManager();
+        $coursesSubscribedToUsergroupIds = $usergroupLib->get_courses_by_usergroup($usergroupId);
+        if (count($coursesSubscribedToUsergroupIds) > 0) {
+            $coursesSubscribedToUsergroup = [];
+            foreach ($coursesSubscribedToUsergroupIds as $courseId) {
+                $coursesSubscribedToUsergroup[] = [
+                    'code' => $courseManager->get_course_code_from_course_id($courseId),
+                    'name' => $courseManager->getCourseNameFromCode($courseManager->get_course_code_from_course_id($courseId)),
+                ];
+            }
+            $data['coursesSubscribedToUsergroup'] = StandardizationService::sort($coursesSubscribedToUsergroup);
+        }
+
+        if (count($usersSubscribedToUsergroupIds) + count($coursesSubscribedToUsergroupIds) == 0) {
+            $data['warning'] = get_lang('No user and no course are subscribed to this class');
+        } else if (count($usersSubscribedToUsergroupIds) == 0) {
+            $data['warning'] = get_lang('No user are subscribed to this class');
+        } else if (count($coursesSubscribedToUsergroupIds) == 0) {
+            $data['warning'] = get_lang('No course are subscribed to this class');
+        }
+
+        return $data;
+    }
+
 
     public static function getRoleName($relation)
     {
